requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv==1.0.1
requests>=2.31.0
drf-yasg>=1.21.7 